+++ src/schema.js
@@ -10,7 +10,7 @@ const blocks = {
 		labels: [null, DEFAULT_LABEL]
 	},
 	backend: {
-		parents: ['beta_oauth2', 'definitions', 'jwt', 'oauth2', 'oidc', 'proxy', 'request'],
+		parents: ['beta_oauth2', 'beta_token_request', 'definitions', 'jwt', 'oauth2', 'oidc', 'proxy', 'request'],
 		description: "Defines the connection to a local/remote backend service.",
 		examples: ['backend-configuration'],
 		labels: (parentBlockName) => {
@@ -25,12 +25,29 @@ const blocks = {
 	beta_oauth2: {
 		parents: ['definitions'],
 		description: "Access control for an OAuth2 Authorization Code Grant Flow redirect endpoint.",
+		docs: '/configuration/block/oauth2',
 		labelled: true
 	},
 	beta_health: {
 		parents: ['backend'],
 		examples: ['health-check'],
+		docs: '/configuration/block/health',
 		labelled: false
+	},
+	beta_rate_limit: {
+		parents: (context) => {
+			if (context.length >= 2 && context[0].name === "backend" && context[1].name === "definitions") {
+				return ['backend']
+			}
+			return `"beta_rate_limit" only valid in a "backend" in a "definitions" block.`
+		},
+		description: "Protects backend services. It implements quota management used to avoid cascading failures or to spare resources.",
+		labelled: false
+	},
+	beta_token_request: {
+		parents: ['backend'],
+		description: "Configures a request to get a token used to authorize backend requests.",
+		labels: [null, DEFAULT_LABEL]
 	},
 	cors: {
 		parents: ['api', 'files', 'server', 'spa'],
@@ -50,6 +67,19 @@ const blocks = {
 		description: "Defines the entry points of Couper.",
 		labels: ['/']
 	},
+	environment: {
+		preprocessed: true,
+		description: "Refines the configuration based on the current environment.",
+		labels: [DEFAULT_LABEL],
+		parents: context => {
+			for (const item of context) {
+				if (item.name === "environment" && item.type === "block") {
+					return `Nested "environment" blocks are not allowed.`
+				}
+			}
+			return ALL_BLOCKS_BUT_ENVIRONMENT.concat([null]) // top-level
+		}
+	},
 	error_handler: {
 		parents: ['api', 'basic_auth', 'beta_oauth2', 'endpoint', 'jwt', 'oidc', 'saml'],
 		examples: ['error-handling-ba', 'sequences'],
@@ -57,9 +87,9 @@ const blocks = {
 			return [null].concat(blocks.error_handler._labelsForParent[parentBlockName])
 		},
 		_labelsForParent: {
-			'api':         ['access_control', 'backend', 'backend_timeout', 'backend_openapi_validation', 'backend_unhealthy', 'beta_insufficient_permissions'],
+			'api':         ['access_control', 'backend', 'backend_timeout', 'backend_openapi_validation', 'backend_unhealthy', 'beta_backend_token_request', 'beta_insufficient_permissions'],
 			'basic_auth':  ['access_control', 'basic_auth', 'basic_auth_credentials_missing'],
-			'endpoint':    ['access_control', 'backend', 'backend_timeout', 'backend_openapi_validation', 'backend_unhealthy', 'beta_insufficient_permissions', 'endpoint', 'sequence', 'unexpected_status'],
+			'endpoint':    ['access_control', 'backend', 'backend_timeout', 'backend_openapi_validation', 'backend_unhealthy', 'beta_backend_token_request', 'beta_insufficient_permissions', 'endpoint', 'sequence', 'unexpected_status'],
 			'jwt':         ['access_control', 'jwt', 'jwt_token_expired', 'jwt_token_invalid', 'jwt_token_missing'],
 			'saml':        ['access_control', 'saml'],
 			'beta_oauth2': ['access_control', 'oauth2'],
@@ -87,6 +117,7 @@ const blocks = {
 	oauth2: {
 		parents: ['backend'],
 		description: "Configures the OAuth2 Client Credentials flow to request a bearer token for its backend request.",
+		docs: '/configuration/block/oauth2req_auth',
 		examples: ['oauth2-client-credentials'],
 		labelled: false
 	},
@@ -195,11 +226,14 @@ const attributes = {
 		examples: ['saml'],
 		type: 'tuple'
 	},
+	assertion: {
+		parents: ['oauth2']
+	},
 	authorization_endpoint: {
 		parents: ['beta_oauth2']
 	},
 	backend: { // label reference
-		parents: ['beta_oauth2', 'jwt', 'oauth2', 'oidc', 'proxy', 'request'],
+		parents: ['beta_oauth2', 'beta_token_request', 'jwt', 'oauth2', 'oidc', 'proxy', 'request'],
 		definingBlocks: ["backend"],
 		examples: ['backend-configuration']
 	},
@@ -244,14 +278,14 @@ const attributes = {
 		parents: ['settings'],
 	},
 	body: {
-		parents: ['request', 'response']
+		parents: ['beta_token_request', 'request', 'response']
 	},
 	bootstrap_file: {
 		parents: ['spa'],
 		examples: ['spa-serving']
 	},
 	ca_file: {
-	  parents: ['settings']
+		parents: ['settings']
 	},
 	claims: {
 		parents: ['jwt', 'jwt_signing_profile'],
@@ -333,13 +367,16 @@ const attributes = {
 		examples: ['simple-fileserving']
 	},
 	expected_status: {
-		parents: ['beta_health', 'proxy', 'request'],
+		parents: ['beta_health', 'beta_token_request', 'proxy', 'request'],
 		type: 'tuple',
 		examples: ['sequences'],
 		tupleType: 'number'
 	},
 	expected_text: {
 		parents: ['beta_health']
+	},
+	environment: {
+		parents: ['settings']
 	},
 	failure_threshold: {
 		parents: ['beta_health'],
@@ -349,19 +386,19 @@ const attributes = {
 		parents: ['openapi']
 	},
 	form_body: {
-		parents: ['request'],
+		parents: ['beta_token_request', 'request'],
 		type: 'object'
 	},
 	grant_type: {
 		parents: ['beta_oauth2', 'oauth2'],
-		options: ['authorization_code', 'client_credentials']
+		options: ['authorization_code', 'client_credentials', 'password', 'urn:ietf:params:oauth:grant-type:jwt-bearer']
 	},
 	header: {
 		parents: ['jwt'],
 		examples: ['jwt-access-control']
 	},
 	headers: {
-		parents: ['beta_health', 'jwt_signing_profile', 'request', 'response'],
+		parents: ['beta_health', 'beta_token_request', 'jwt_signing_profile', 'request', 'response'],
 		examples: ['static-responses'],
 		type: 'object'
 	},
@@ -403,7 +440,7 @@ const attributes = {
 		type: 'duration'
 	},
 	json_body: {
-		parents: ['request', 'response'],
+		parents: ['beta_token_request', 'request', 'response'],
 		examples: ['static-responses'],
 		type: ['boolean', 'number', 'string', 'object', 'tuple'],
 	},
@@ -452,7 +489,11 @@ const attributes = {
 		type: 'number'
 	},
 	method: {
-		parents: ['request']
+		parents: ['beta_token_request', 'request']
+	},
+	mode: {
+		parents: ['beta_rate_limit'],
+		options: ['wait', 'block']
 	},
 	no_proxy_from_env: {
 		parents: ['settings'],
@@ -462,7 +503,7 @@ const attributes = {
 		parents: ['backend']
 	},
 	password: {
-		parents: ['basic_auth']
+		parents: ['basic_auth', 'oauth2']
 	},
 	path: {
 		parents: ['backend', 'beta_health']
@@ -475,11 +516,23 @@ const attributes = {
 		examples: ['spa-serving'],
 		type: 'tuple'
 	},
+	period: {
+		parents: ['beta_rate_limit'],
+		type: 'duration'
+	},
+	period_window: {
+		parents: ['beta_rate_limit'],
+		options: ['sliding', 'fixed']
+	},
+	per_period: {
+		parents: ['beta_rate_limit'],
+		type: 'number'
+	},
 	proxy: {
 		parents: ['backend']
 	},
 	query_params: {
-		parents: ['request'],
+		parents: ['beta_token_request', 'request'],
 		type: 'object'
 	},
 	realm: {
@@ -603,6 +656,9 @@ const attributes = {
 		parents: ['beta_oauth2', 'oauth2', 'oidc'],
 		options: ['client_secret_basic', 'client_secret_post']
 	},
+	token: {
+		parents: ['beta_token_request']
+	},
 	token_value: {
 		parents: ['jwt'],
 		examples: ['jwt-access-control'],
@@ -613,11 +669,11 @@ const attributes = {
 		type: 'duration'
 	},
 	ttl: {
-		parents: ['jwt_signing_profile'],
+		parents: ['beta_token_request', 'jwt_signing_profile'],
 		type: 'duration'
 	},
 	url: {
-		parents: ['request', 'proxy']
+		parents: ['beta_token_request', 'request', 'proxy']
 	},
 	use_when_unhealthy: {
 		parents: ['backend'],
@@ -625,6 +681,9 @@ const attributes = {
 	},
 	user: {
 		parents: ['basic_auth']
+	},
+	username: {
+		parents: ['oauth2']
 	},
 	userinfo_backend: { // label reference
 		parents: ['oidc'],
@@ -686,6 +745,24 @@ const functions = {
 const commonProperties = ['body', 'context', 'cookies', 'headers', 'json_body']
 
 const variables = {
+	backend: {
+		parents: ['backend'],
+		description: "An object with backend attributes.",
+		values: [
+			'health', // TODO how to add health object properties?
+			'beta_tokens',
+			'beta_token'
+		]
+	},
+	backends: {
+		child: 'default',
+		description: "An object with all backends and their attributes. To access a specific backend use the related name.",
+		values: [
+			'health',
+			'beta_tokens',
+			'beta_token'
+		]
+	},
 	backend_request: {
 		parents: ['backend'],
 		description: "Holds information about the current backend request.",
@@ -731,7 +808,7 @@ const variables = {
 		])
 	},
 	couper: {
-		values: ['version']
+		values: ['environment', 'version']
 	},
 	env: {
 		description: "The value of an environment variable.",
@@ -752,7 +829,17 @@ const variables = {
 			'query',
 			'url'
 		])
+	},
+	beta_token_response: {
+		parents: ['beta_token_request'],
+		description: "Holds information about the current token response.",
+		values: commonProperties.concat([
+			'status'
+		])
 	}
 }
+
+const ALL_BLOCKS = Object.keys(blocks)
+const ALL_BLOCKS_BUT_ENVIRONMENT = ALL_BLOCKS.filter(block => block !== "environment")
 
 module.exports = { attributes, blocks, functions, variables, DEFAULT_LABEL }